Refactor session tests to use concrete cache class
Replaced instances of the interface `IApplicationUserSessionCache` with the concrete class `ApplicationUserSessionCache` in multiple test methods within `SessionTests.cs`. This change applies to tests for setting and retrieving items using both type keys and string keys, as well as for value types. The update aims to simplify the tests and leverage specific features of the `ApplicationUserSessionCache` class.