Add deploy job to build-dev-beta.yml
This commit introduces a new `deploy` job in the `build-dev-beta.yml` file. The job runs on a self-hosted Linux runner with write permissions and depends on the `build` job. It includes steps to download the latest `blazam-dev-beta.zip` release artifact from the `BetaDev1` tag and extract its contents to the `/opt/blazam` directory.