Add FQDN link rendering in EmailFooter component
Introduces a new using directive for `BLAZAM.Helpers` and adds a conditional block to construct a URL based on the application's FQDN. It determines the protocol (HTTP or HTTPS) based on listening addresses and displays a link with the application name using `MudText` and `MudLink`. Removed the commented-out `EmailSiteNavigation` line.