Add support for Description field in OrganizationSection
A new condition was added to the `@if` statement in the
`OrganizationSection.razor` file to check if the
`GroupableEntry.CanReadField(ActiveDirectoryFields.Description)`
returns true. This enhances the component by including the
`Description` field for evaluation alongside other Active
Directory fields such as `EmployeeId`, `Department`, `Company`,
`Manager`, `Title`, and `PhysicalDeliveryOffice`.

Fixes #1278